refactor: theme관련 코드 분리
